remove abstractions to go direct from workbench to openai
