+++ assistants/codespace-assistant/assistant/response/models.py
@@ -1,0 +1,10 @@
+from typing import Any, Literal
+
+from attr import dataclass
+
+
+@dataclass
+class StepResult:
+    status: Literal["final", "error", "continue"]
+    conversation_tokens: int = 0
+    metadata: dict[str, Any] | None = None
