--- assistants/codespace-assistant/assistant/response/providers/__init__.py
@@ -1,10 +1,0 @@
-from .anthropic_response_provider import AnthropicResponseProvider
-from .base_provider import NumberTokensResult, ResponseProvider
-from .openai_response_provider import OpenAIResponseProvider
-
-__all__ = [
-    "AnthropicResponseProvider",
-    "NumberTokensResult",
-    "OpenAIResponseProvider",
-    "ResponseProvider",
-]
