+++ assistants/codespace-assistant/assistant/response/utils/__init__.py
@@ -1,20 +1,18 @@
 from .formatting_utils import get_response_duration_message, get_token_usage_message
 from .message_utils import (
     build_system_message_content,
-    conversation_message_to_completion_messages,
+    conversation_message_to_chat_message_params,
     get_history_messages,
-    inject_attachments_inline,
 )
-from .response_provider_utils import initialize_response_provider
-from .token_utils import num_tokens_from_messages
+from .openai_utils import convert_mcp_tools_to_openai_tools, extract_content_from_tool_calls, get_completion
 
 __all__ = [
     "build_system_message_content",
-    "conversation_message_to_completion_messages",
+    "convert_mcp_tools_to_openai_tools",
+    "conversation_message_to_chat_message_params",
+    "extract_content_from_tool_calls",
+    "get_completion",
     "get_history_messages",
     "get_response_duration_message",
     "get_token_usage_message",
-    "initialize_response_provider",
-    "inject_attachments_inline",
-    "num_tokens_from_messages",
 ]
