+++ assistants/codespace-assistant/assistant/response/utils/message_utils.py
@@ -1,9 +1,16 @@
 import json
 import logging
-import re
-from typing import Any, Awaitable, Callable, Sequence
+from typing import Any
 
-from assistant_extensions.ai_clients.model import CompletionMessage
+import openai_client
+from openai.types.chat import (
+    ChatCompletionAssistantMessageParam,
+    ChatCompletionMessageParam,
+    ChatCompletionMessageToolCallParam,
+    ChatCompletionSystemMessageParam,
+    ChatCompletionToolMessageParam,
+    ChatCompletionUserMessageParam,
+)
 from semantic_workbench_api_model.workbench_model import (
     ConversationMessage,
     ConversationParticipant,
@@ -11,11 +18,8 @@ from semantic_workbench_api_model.workbench_model import (
 )
 from semantic_workbench_assistant.assistant_app import ConversationContext
 
-from assistant.config import AssistantConfigModel
-from assistant.response.utils.formatting_utils import format_message
-
-from ..providers import ResponseProvider
-from .token_utils import num_tokens_from_messages
+from ...config import AssistantConfigModel
+from .formatting_utils import format_message
 
 logger = logging.getLogger(__name__)
 
@@ -53,85 +57,131 @@ def build_system_message_content(
     return system_message_content
 
 
-async def conversation_message_to_completion_messages(
-    context: ConversationContext, message: ConversationMessage, participants: list[ConversationParticipant]
-) -> list[CompletionMessage]:
+def conversation_message_to_tool_message(
+    message: ConversationMessage,
+) -> ChatCompletionToolMessageParam | None:
     """
-    Convert a conversation message to a list of completion messages.
+    Convert a conversation message to a tool message.
+    """
+    tool_result = message.metadata.get("tool_result")
+    if tool_result is not None:
+        content = tool_result.get("content")
+        tool_call_id = tool_result.get("tool_call_id")
+        if content is not None and tool_call_id is not None:
+            return ChatCompletionToolMessageParam(
+                role="tool",
+                content=content,
+                tool_call_id=tool_call_id,
+            )
+
+
+def tool_calls_from_metadata(metadata: dict[str, Any]) -> list[ChatCompletionMessageToolCallParam] | None:
+    """
+    Get the tool calls from the message metadata.
+    """
+    if metadata is None or "tool_calls" not in metadata:
+        return None
+
+    tool_calls = metadata["tool_calls"]
+    if not isinstance(tool_calls, list) or len(tool_calls) == 0:
+        return None
+
+    tool_call_params: list[ChatCompletionMessageToolCallParam] = []
+    for tool_call in tool_calls:
+        id = tool_call["id"]
+        name = tool_call["name"]
+        arguments = json.dumps(tool_call["arguments"])
+        if id is not None and name is not None and arguments is not None:
+            tool_call_params.append(
+                ChatCompletionMessageToolCallParam(
+                    id=id,
+                    type="function",
+                    function={"name": name, "arguments": arguments},
+                )
+            )
+
+    return tool_call_params
+
+
+def conversation_message_to_assistant_message(
+    message: ConversationMessage,
+    participants: list[ConversationParticipant],
+) -> ChatCompletionAssistantMessageParam:
+    """
+    Convert a conversation message to an assistant message.
+    """
+    assistant_message = ChatCompletionAssistantMessageParam(
+        role="assistant",
+        content=format_message(message, participants),
+    )
+
+    # get the tool calls from the message metadata
+    tool_calls = tool_calls_from_metadata(message.metadata)
+    if tool_calls:
+        assistant_message["tool_calls"] = tool_calls
+
+    return assistant_message
+
+
+def conversation_message_to_user_message(
+    message: ConversationMessage,
+    participants: list[ConversationParticipant],
+) -> ChatCompletionMessageParam:
+    """
+    Convert a conversation message to a user message.
+    """
+    return ChatCompletionUserMessageParam(
+        role="user",
+        content=format_message(message, participants),
+    )
+
+
+async def conversation_message_to_chat_message_params(
+    context: ConversationContext, message: ConversationMessage, participants: list[ConversationParticipant]
+) -> list[ChatCompletionMessageParam]:
+    """
+    Convert a conversation message to a list of chat message parameters.
     """
 
     # some messages may have multiple parts, such as a text message with an attachment
-    completion_messages: list[CompletionMessage] = []
+    chat_message_params: list[ChatCompletionMessageParam] = []
 
-    # add the message to the completion messages, treating any message from a source other than the assistant
-    # as a user message
+    # add the message to list, treating messages from a source other than this assistant as a user message
     if message.message_type == MessageType.tool_result:
-        tool_result = message.metadata.get("tool_result")
-        if tool_result:
-            completion_messages.append(
-                CompletionMessage(
-                    role="tool",
-                    content=tool_result.get("content"),
-                    tool_call_id=tool_result.get("tool_call_id"),
-                    metadata=message.metadata,
-                )
-            )
+        tool_message = conversation_message_to_tool_message(message)
+        if tool_message is not None:
+            chat_message_params.append(tool_message)
+        else:
+            logger.warning(f"Failed to convert tool message to completion message: {message}")
 
     elif message.sender.participant_id == context.assistant.id:
-        # get the tool calls from the message metadata
-        tool_calls: list[dict[str, Any]] | None = None
-        if message.metadata is not None and "tool_calls" in message.metadata:
-            # FIXME: this is forcing the OpenAI format, but it should be more generic
-            tool_calls = [
-                {
-                    "id": tool_call["id"],
-                    "type": "function",
-                    "function": {"name": tool_call["name"], "arguments": json.dumps(tool_call["arguments"])},
-                }
-                for tool_call in message.metadata["tool_calls"]
-            ]
-
-        completion_messages.append(
-            CompletionMessage(
-                role="assistant",
-                content=format_message(message, participants),
-                tool_calls=tool_calls,
-                metadata=message.metadata,
-            )
-        )
+        # add the assistant message to the completion messages
+        assistant_message = conversation_message_to_assistant_message(message, participants)
+        chat_message_params.append(assistant_message)
 
     else:
         # add the user message to the completion messages
-        completion_messages.append(
-            CompletionMessage(
-                role="user",
-                content=format_message(message, participants),
-                metadata=message.metadata,
-            )
-        )
+        user_message = conversation_message_to_user_message(message, participants)
+        chat_message_params.append(user_message)
 
+        # add the attachment message to the completion messages
         if message.filenames and len(message.filenames) > 0:
             # add a system message to indicate the attachments
-            completion_messages.append(
-                CompletionMessage(
-                    role="system", content=f"Attachment(s): {', '.join(message.filenames)}", metadata=message.metadata
+            chat_message_params.append(
+                ChatCompletionSystemMessageParam(
+                    role="system", content=f"Attachment(s): {', '.join(message.filenames)}"
                 )
             )
 
-    return completion_messages
+    return chat_message_params
 
 
 async def get_history_messages(
-    response_provider: ResponseProvider,
     context: ConversationContext,
     participants: list[ConversationParticipant],
-    converter: Callable[
-        [ConversationContext, ConversationMessage, list[ConversationParticipant]],
-        Awaitable[list[CompletionMessage]],
-    ],
     model: str,
     token_limit: int | None = None,
-) -> list[CompletionMessage]:
+) -> list[ChatCompletionMessageParam]:
     """
     Get all messages in the conversation, formatted for use in a completion.
     """
@@ -161,20 +211,8 @@ async def get_history_messages(
         # messages are returned in reverse order, so we need to reverse them
         for message in reversed(messages_list):
             # format the message
-            formatted_message_list = await converter(context, message, participants)
-            try:
-                results = await num_tokens_from_messages(
-                    context=context,
-                    response_provider=response_provider,
-                    messages=formatted_message_list,
-                    model=model,
-                    metadata={},
-                    metadata_key="get_history_messages",
-                )
-                if results is not None:
-                    token_count += results.count
-            except Exception as e:
-                logger.exception(f"exception occurred calculating token count: {e}")
+            formatted_message_list = await conversation_message_to_chat_message_params(context, message, participants)
+            token_count += openai_client.num_tokens_from_messages(formatted_message_list, model=model)
 
             # if a token limit is provided and the token count exceeds the limit, break the loop
             if token_limit and token_count > token_limit:
@@ -185,56 +223,3 @@ async def get_history_messages(
 
     # return the formatted messages
     return history
-
-
-def inject_attachments_inline(
-    history_messages: list[CompletionMessage],
-    attachment_messages: Sequence[CompletionMessage],
-) -> list[CompletionMessage]:
-    """
-    Inject the attachment messages inline into the history messages.
-    """
-
-    # iterate over the history messages and for every message that contains an attachment,
-    # find the related attachment message and replace the attachment message with the inline attachment content
-    for index, history_message in enumerate(history_messages):
-        # if the history message does not contain content, as a string value, skip
-        content = history_message.content
-        if not content or not isinstance(content, str):
-            # TODO: handle list content, which may contain multiple parts including images
-            continue
-
-        # get the attachment filenames string from the history message content
-        attachment_filenames_string = re.findall(r"Attachment\(s\): (.+)", content)
-
-        # if the history message does not contain an attachment filenames string, skip
-        if not attachment_filenames_string:
-            continue
-
-        # split the attachment filenames string into a list of attachment filenames
-        attachment_filenames = [filename.strip() for filename in attachment_filenames_string[0].split(",")]
-
-        # initialize a list to store the replacement messages
-        replacement_messages = []
-
-        # iterate over the attachment filenames and find the related attachment message
-        for attachment_filename in attachment_filenames:
-            # find the related attachment message
-            attachment_message = next(
-                (
-                    attachment_message
-                    for attachment_message in attachment_messages
-                    if f"<ATTACHMENT><FILENAME>{attachment_filename}</FILENAME>" in str(attachment_message.content)
-                ),
-                None,
-            )
-
-            if attachment_message:
-                # replace the attachment message with the inline attachment content
-                replacement_messages.append(attachment_message)
-
-        # if there are replacement messages, replace the history message with the replacement messages
-        if len(replacement_messages) > 0:
-            history_messages[index : index + 1] = replacement_messages
-
-    return history_messages
